+++ src/hw_cpu/v810/v810_fp_ops.h
@@ -64,7 +64,7 @@ class V810_FP_Ops
  bool fp_is_zero(uint32 v);
  bool fp_is_inf_nan_sub(uint32 v);
 
- uint8 clz64(uint64 v);
+ unsigned clz64(uint64 v);
  void fpim_decode(fpim* df, uint32 v);
  void fpim_round(fpim* df);
  void fpim_round_int(fpim* df, bool truncate = false);
